+++ backend/src/neuroagent/agent_routine.py
@@ -354,7 +354,6 @@ class AgentsRoutine:
                 # Stage the new message for addition to DB
                 messages.append(
                     Messages(
-                        order=len(messages),
                         thread_id=messages[-1].thread_id,
                         entity=get_entity(message),
                         content=json.dumps(message),
@@ -416,7 +415,6 @@ class AgentsRoutine:
                 messages.extend(
                     [
                         Messages(
-                            order=len(messages) + i,
                             thread_id=messages[-1].thread_id,
                             entity=Entity.TOOL,
                             content=json.dumps(tool_response),
@@ -481,7 +479,6 @@ class AgentsRoutine:
             ):
                 messages.append(
                     Messages(
-                        order=len(messages),
                         thread_id=messages[-1].thread_id,
                         entity=get_entity(message),
                         content=json.dumps(message),
@@ -495,7 +492,6 @@ class AgentsRoutine:
                 messages.extend(
                     [
                         Messages(
-                            order=len(messages),
                             thread_id=messages[-1].thread_id,
                             entity=Entity.TOOL,
                             content=json.dumps(