+++ backend/src/neuroagent/app/routers/qa.py
@@ -193,7 +193,6 @@ async def stream_chat_agent(
     ):
         messages.append(
             Messages(
-                order=len(messages),
                 thread_id=thread.thread_id,
                 entity=Entity.USER,
                 content=json.dumps({"role": "user", "content": user_request.content}),
@@ -229,7 +228,6 @@ async def stream_chat_agent(
                 )
                 messages.append(
                     Messages(
-                        order=len(messages),
                         thread_id=thread.thread_id,
                         entity=Entity.AI_MESSAGE,
                         content=json.dumps({"role": "assistant", "content": response}),