+++ backend/src/neuroagent/app/schemas.py
@@ -23,7 +23,6 @@ class MessageResponse(BaseModel):
     message_id: str
     entity: str
     thread_id: str
-    order: int
     is_complete: bool
     creation_date: datetime.datetime
     msg_content: dict[str, Any]
@@ -72,7 +71,6 @@ class MessagesRead(BaseModel):
     """Output of the conversation listing crud."""
 
     message_id: str
-    order: int
     creation_date: datetime.datetime
     msg_content: str
     entity: Literal[Entity.USER, Entity.AI_MESSAGE]